Fixed warnings
git-svn-id: http://www.geogebra.org/svn/trunk/geogebra@15032 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/geogebra/common/io/MyXMLHandler.java
@@ -1610,8 +1610,8 @@ public class MyXMLHandler implements DocHandler {
 		geogebra.common.awt.Dimension evSize = app.getSettings().getEuclidian(1).getPreferredSize();
 		
 		// calculate window dimensions
-		int width = (int) evSize.getWidth();
-		int height = (int) evSize.getHeight();
+		int width = evSize.getWidth();
+		int height = evSize.getHeight();
 		
 		// minimal size for documents, necessary for GeoGebra < 3
 		if(width <= 100 || height <= 100) {
